requirement.txt
django
djangorestframework
django-cors-headers
celery 
rabbitmq 
drf-yasg
celery==5.3.6
django-celery-results
